# docs/openapi/openai-compat.yaml
openapi: 3.1.0
info:
  title: Tama — OpenAI-Compatible API & Root Endpoints
  description: |
    This file documents two groups of endpoints served by Tama:

    1. **OpenAI-compatible `/v1/` routes** — Chat completions and model listing.
       These follow the OpenAI API spec closely so that any client written for
       OpenAI can be pointed at Tama with minimal changes.

    2. **Root utility endpoints** — `/health`, `/status`, and `/metrics`.
       `/metrics` returns Prometheus exposition format (merged proxy + backend
       metrics). For richer system health, use `GET /tama/v1/system/health`
       (see `tama-api.yaml`).

    **Base URL:** Point your OpenAI client at `http://localhost:8080` and set
    `OPENAI_API_KEY` to a `tama_` API key (or any non-empty string when API
    keys are disabled).
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local Tama instance (default port)
...
paths:
...
  /v1/chat/completions:
    post:
      operationId: createChatCompletion
      summary: Create a chat completion
      description: |
        Proxies the request to the appropriate backend for the requested model.

        If the model is not currently loaded, Tama will automatically start its
        backend before forwarding the request. Startup time is bounded by
        `proxy.startup_timeout_secs` in the Tama config.

        Supports both regular (buffered) and streaming (`"stream": true`) responses.
        For streaming, prefer `POST /v1/chat/completions/stream` to avoid
        buffering the full response body.

        **Model field:** The `model` field must match a key in the Tama config
        (e.g. `llama3`) or the underlying model name that config entry points to.
      tags: [chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
            examples:
              basic:
                summary: Simple user message
                value:
                  model: llama3
                  messages:
                    - role: user
                      content: "What is the capital of France?"
              with_system:
                summary: System + user message
                value:
                  model: llama3
                  messages:
                    - role: system
                      content: "You are a helpful assistant."
                    - role: user
                      content: "Explain quantum entanglement simply."
                  temperature: 0.7
                  max_tokens: 512
              streaming:
                summary: Streaming response
                value:
                  model: llama3
                  messages:
                    - role: user
                      content: "Write a short poem about Rust."
                  stream: true
      responses:
        "200":
          description: |
            Chat completion response. When `stream` is `false` (default), the
            full completion is returned as JSON. When `stream` is `true`, the
            response is a sequence of `text/event-stream` SSE chunks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
              example:
                id: chatcmpl-abc123
                object: chat.completion
                created: 1712000000
                model: llama3
                choices:
                  - index: 0
                    message:
                      role: assistant
                      content: "The capital of France is Paris."
                    finish_reason: stop
                usage:
                  prompt_tokens: 14
                  completion_tokens: 9
                  total_tokens: 23
            text/event-stream:
              schema:
                type: string
                description: |
                  Server-Sent Events stream. Each event is a `data:` line
                  containing a `ChatCompletionChunk` JSON object. The stream
                  ends with `data: [DONE]`.
                example: |
                  data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1712000000,"model":"llama3","choices":[{"index":0,"delta":{"role":"assistant","content":"The"},"finish_reason":null}]}

                  data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1712000000,"model":"llama3","choices":[{"index":0,"delta":{"content":" capital"},"finish_reason":null}]}

                  data: [DONE]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/chat/completions/stream:
    post:
      operationId: createStreamingChatCompletion
      summary: Create a streaming chat completion
      description: |
        Identical to `POST /v1/chat/completions` with `"stream": true`, but
        provided as a dedicated endpoint so that clients can target it directly
        without needing to set the stream flag in the body.

        The response is always a `text/event-stream` SSE stream regardless of
        the `stream` field in the request body.
      tags: [chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
            example:
              model: llama3
              messages:
                - role: user
                  content: "Count to five."
      responses:
        "200":
          description: SSE stream of chat completion chunks
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Server-Sent Events stream. Each event is a `data:` line
                  containing a `ChatCompletionChunk` JSON object. Ends with
                  `data: [DONE]`.
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ===========================================================================
  # OpenAI-compatible: Models
  # ===========================================================================
...
  /v1/models/{model_id}:
    get:
      operationId: getModel
      summary: Get a model by ID
      description: |
        Returns a single model object for the given model ID.

        The `model_id` can be either the Tama config key (e.g. `llama3`) or
        the underlying model name (e.g. `meta-llama/Meta-Llama-3-8B-Instruct`).

        Returns the model's `ready` state (Tama extension) in addition to
        standard OpenAI fields.
      tags: [models]
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
            example: llama3
          description: Model config key or underlying model name
      responses:
        "200":
          description: Model object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
              example:
                id: llama3
                object: model
                created: 1712000000
                owned_by: llama-cpp
                ready: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/model/info:
    get:
      operationId: getLiteLLMModelInfo
      summary: List model info (LiteLLM-compatible discovery)
      description: |
        Returns all enabled models (plus alias entries) in LiteLLM's
        `model_info_v1` discovery shape: `{ "data": [ { model_name,
        litellm_params, model_info } ] }` — one entry per enabled model plus
        alias entries (same set and inheritance rules as the opencode
        discovery endpoint). `?litellm_model_id=<name>` filters to the single
        entry whose `model_name` matches (an unknown id yields an empty
        `data` array; an absent or empty value returns the full list).

        Requires the `Inference` scope. The same handler is served at
        `GET /model/info` (also requires the `Inference` scope).
      tags: [models]
      parameters:
        - name: litellm_model_id
          in: query
          required: false
          schema:
            type: string
            example: qwen3-8b
          description: |
            Filter to the single entry whose `model_name` matches. Unknown id
            → empty `data`; absent/empty → full list.
      responses:
        "200":
          description: Model info entries in the LiteLLM `data` array shape
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiteLLMModelInfoList"
              example:
                data:
                  - model_name: qwen3-8b
                    litellm_params:
                      model: qwen3-8b
                      provider: llama.cpp
                    model_info:
                      id: qwen3-8b
                      mode: chat
                      litellm_provider: llama.cpp
                      max_input_tokens: 262144
                      max_output_tokens: 32768
                      max_tokens: 32768
                      supports_function_calling: true
                      supports_reasoning: true
                      supports_vision: true
                      supports_audio_input: false
                      supports_pdf_input: false
                      input_cost_per_token: 0.0
                      output_cost_per_token: 0.0
                      cache_read_input_token_cost: 0.0
                      cache_creation_input_token_cost: 0.0
                      reasoning_effort_levels: [none, low, medium, xhigh]
                      supports_none_reasoning_effort: true
                      supports_minimal_reasoning_effort: false
                      supports_low_reasoning_effort: true
                      supports_xhigh_reasoning_effort: true
                      supports_max_reasoning_effort: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /model/info:
    get:
      operationId: getLiteLLMModelInfoAlt
      summary: List model info (LiteLLM-compatible)
      description: |
        Identical handler and response to `GET /v1/model/info`, served at the
        non-prefixed path. Requires a valid API key (auth middleware) and the
        `Inference` scope — the same authorization contract as its `/v1` twin.
      tags: [models]
      parameters:
        - name: litellm_model_id
          in: query
          required: false
          schema:
            type: string
            example: qwen3-8b
          description: |
            Filter to the single entry whose `model_name` matches. Unknown id
            → empty `data`; absent/empty → full list.
      responses:
        "200":
          description: Model info entries in the LiteLLM `data` array shape
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiteLLMModelInfoList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ===========================================================================
  # Root utility endpoints
  # ===========================================================================
...
components:
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              message: "Model 'foo' not found"
              type: NotFoundError
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              message: "Missing required field: model"
              type: BadRequestError
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              message: "Failed to load model: timeout after 60s"
              type: LoadModelError
    Unauthorized:
      description: "Missing or invalid API key (or a `tama_` key provided while API keys are disabled)"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              message: "Invalid or missing API key"
              type: UnauthorizedError
    Forbidden:
      description: "API key lacks the required scope for this route (e.g. `Inference` for `/v1/*`)"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              message: "missing required scope: inference"
              type: ForbiddenError
  schemas:
...
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: |
            Tama config key (e.g. `llama3`) or underlying model name.
            If the model is not loaded, Tama will start its backend automatically.
          example: llama3
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ChatMessage"
          description: Conversation history including the new user turn
        temperature:
          type: number
          format: float
          minimum: 0
          maximum: 2
          description: Sampling temperature. Higher values produce more random output.
          example: 0.7
        top_p:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Nucleus sampling probability mass
          example: 0.9
        max_tokens:
          type: integer
          format: int32
          minimum: 1
          description: Maximum number of tokens to generate
          example: 512
        stream:
          type: boolean
          default: false
          description: Whether to stream the response as Server-Sent Events
          example: false
        stop:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Stop sequence(s). Generation halts when any of these strings is produced.
          example: ["\n\n", "Human:"]
        presence_penalty:
          type: number
          format: float
          minimum: -2
          maximum: 2
          description: Penalise new tokens based on whether they appear in the text so far
          example: 0.0
        frequency_penalty:
          type: number
          format: float
          minimum: -2
          maximum: 2
          description: Penalise new tokens based on their frequency in the text so far
          example: 0.0
        user:
          type: string
          description: Unique identifier for the end user (passed through to backend)
...
    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
          description: Unique completion ID
          example: chatcmpl-abc123
        object:
          type: string
          enum: [chat.completion]
          example: chat.completion
        created:
          type: integer
          format: int64
          description: Unix timestamp of creation
          example: 1712000000
        model:
          type: string
          description: Model that generated the completion
          example: llama3
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionChoice"
        usage:
          $ref: "#/components/schemas/UsageStats"
          description: Token usage statistics (may be absent in streaming)

    # -------------------------------------------------------------------------
    # Model listing schemas (OpenAI-compatible)
    # -------------------------------------------------------------------------
    Model:
      type: object
      required: [id, object, created, owned_by]
      properties:
        id:
          type: string
          description: Model identifier (Tama config key)
          example: llama3
        object:
          type: string
          enum: [model]
          example: model
        created:
          type: integer
          format: int64
          description: Unix timestamp when the model was loaded (0 if never loaded)
          example: 1712000000
        owned_by:
          type: string
          description: Backend that owns/serves this model
          example: llama-cpp
        ready:
          type: boolean
          description: |
            **Tama extension** (not in OpenAI spec). Whether the backend is
            currently running and accepting requests.
          example: true
...
    LiteLLMModelInfoList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          description: "One entry per enabled model plus alias entries; empty when `?litellm_model_id` matches nothing"
          items:
            $ref: "#/components/schemas/LiteLLMModelInfoEntry"

    # -------------------------------------------------------------------------
    # Root utility schemas
    # -------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"